[example-hybrid] Connect Electron to a server
Add a new example package serving both web browsers and electron clients.
Add new commands for electron to connect to remote servers.

`electron.remote.connect` command to open a QuickOpenMenu and input an URL.
`electron.remote.history.clear` command to clear the local URL history.
`electron.remote.disconnect` to close the window in case you are connected to a server.

Adds an endpoint to tell if it is a Theia application or not.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>